database fcns for user

+++ app/src/main/java/com/cs220/ssmessaging/clientBackend/Message.kt
@@ -16,7 +16,7 @@ interface Message {
     val conversationId : String
     val sender : User
     val recipient : User
-    val timestamp : Int // This should be epoch number
+    val timestamp : Long // This should be epoch number
     fun mEquals(m: Message) : Boolean // decides whether the current message equals another message
     companion object{
         // Need to unit test
@@ -27,7 +27,7 @@ interface Message {
 }
 
 class EncryptedMessage(_message : ByteArray, _conversationId : String,
-                       _messageType : String, _sender : User, _recipient : User, _timestamp : Int) : Message {
+                       _messageType : String, _sender : User, _recipient : User, _timestamp : Long) : Message {
 
     companion object{
         // Need to unit test
@@ -95,7 +95,7 @@ interface UnencryptedMessage : Message {
 }
 
 class TextMessage(_message : String, _conversationId : String,
-                  _sender : User, _recipient : User, _timestamp : Int) : UnencryptedMessage {
+                  _sender : User, _recipient : User, _timestamp : Long) : UnencryptedMessage {
 
     companion object {
         // Need to unit test
@@ -152,7 +152,7 @@ class TextMessage(_message : String, _conversationId : String,
 }
 
 class ImageMessage(_message : ByteArray, _conversationId : String,
-                   _sender : User, _recipient: User, _timestamp : Int) : UnencryptedMessage {
+                   _sender : User, _recipient: User, _timestamp : Long) : UnencryptedMessage {
 
     companion object{
         // Need to unit test
